requirements/local.txt
asgiref==3.2.7
Django==3.0.4
Pillow==10.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1
django-model-utils==4.3.1
django-ckeditor==6.6.1